Refactor out common functions
